Separate manifest resource schemas from runtime config
